server/requirements.txt
websockets>=16.0
python-uinput>=1.0.1

